Notify any listeners of state changes
Create an enumeration of tunnel state change events.  When the tunnel
state machine is started, return a channel that receives state changes
events. Send a state change event before each new state starts handling
events. If a tunnel request is received, the state changes, even if it
returns to the same state. This allows state change events to be sent
every time a request is made.